+++ docs/conf.py
@@ -126,8 +126,10 @@ pygments_style = 'sphinx'
 # a list of builtin themes.
 html_theme = "alabaster"
 html_theme_options = {
-    "font_family": "Palatino, Georgia, serif",
+    "font_family": '"Avenir Next", Calibri, "PT Sans", sans-serif',
+    "head_font_family": '"Avenir Next", Calibri, "PT Sans", sans-serif',
     "font_size": "18px",
+    "page_width": "980px",
 }
 html_logo = "_static/structlog_logo_small.png"