+++ src/structlog/processors.py
@@ -17,9 +17,7 @@ import time
 import six
 
 from structlog._frames import (
-    _find_first_app_frame_and_name,
-    _format_exception,
-    _format_stack,
+    _find_first_app_frame_and_name, _format_exception, _format_stack
 )