+++ tests/test_twisted.py
@@ -12,24 +12,18 @@ import pytest
 
 from pretend import call_recorder
 from six import PY3
-from six.moves import cStringIO as StringIO
-from twisted.python.failure import Failure, NoCurrentExceptionError
-from twisted.python.log import ILogObserver
 
-from structlog._config import _CONFIG
+from six.moves import cStringIO as StringIO
 from structlog import ReturnLogger
+from structlog._config import _CONFIG
 from structlog.processors import KeyValueRenderer
 from structlog.twisted import (
-    BoundLogger,
-    EventAdapter,
-    JSONLogObserverWrapper,
-    JSONRenderer,
-    LoggerFactory,
-    PlainFileLogObserver,
-    ReprWrapper,
-    _extractStuffAndWhy,
-    plainJSONStdOutLogger,
+    BoundLogger, EventAdapter, JSONLogObserverWrapper, JSONRenderer,
+    LoggerFactory, PlainFileLogObserver, ReprWrapper, _extractStuffAndWhy,
+    plainJSONStdOutLogger
 )
+from twisted.python.failure import Failure, NoCurrentExceptionError
+from twisted.python.log import ILogObserver
 
 
 def test_LoggerFactory():